backend/requirements.txt
flask-cors
Werkzeug
numpy==1.24.3
python-dotenv==0.21.0
Flask
Flask-Migrate==3.1.0
Flask
requests
neo4j
jsonify


